--- components/vc-cascader/index.js
@@ -1,3 +1,0 @@
-// based on rc-cascader 0.17.4
-import Cascader from './Cascader';
-export default Cascader;
